fix: refactor struct
