--- packages/runtime/plugin-runtime/src/ssr/serverRender/renderToString/type.ts
@@ -1,14 +1,0 @@
-export enum RenderLevel {
-  CLIENT_RENDER,
-  SERVER_PREFETCH,
-  SERVER_RENDER,
-}
-
-export type RenderResult = {
-  renderLevel: RenderLevel;
-  html?: string;
-  chunksMap: {
-    js: string;
-    css: string;
-  };
-};
